Implement topic bump and others
1) Request tag is now only stylized in request sections
2) Bugfix. Firefox not respecting @@username tab completion
3) Add Your topics to quicklinks
4) Chicken bot Easter Egg random event
5) Option for showing topics with most replies
6) Implement Topic bumping
Rules:
Topic bumping is allowed only in listings sections.
Listing section is defined by the root fid in ACP->general settings

In order to bump a topic, first it must be enabled.

A topic bump can be enabled by:
1) Moderator
2) OP && op in allowed_group && topic_bump!=disabled (by mod)

A topic bump can be disabled by:
1) Moderator

Once the topic bump is enabled the topic can be bumped
1) Moderator can bump at any time
2) The OP of the topic can bump once in a cycle defined in ACP

Disabling and enabling topic bump will reset cycle time.

Any topic in listings can be enabled for bumping by moderator,
even if the topic OP is not in the allowed group.