Implement thanks notification deletion
The problem with thanks notifications is that
notification toggle is attached to the post itself
at the time of the post.
So if the thanks notification was enabled when a
post was made, the poster will continue to get
notfications even if it is turned off later.
The easiest way to deal with this is to simply
delete the notification entry.
To that end, we implement a method to delete all
thanks notificaitons.