Update version number to v0.16.0
Summary for commit

From 2e2131e84face13aa4a2a0bf401c9
*** Implement group based search ***

From c0da1f0f3886daf2c
*** Implement thanks notification deletion ***

The problem with thanks notifications is that
notification toggle is attached to the post itself
at the time of the post.
So if the thanks notification was enabled when a
post was made, the poster will continue to get
notfications even if it is turned off later.
The easiest way to deal with this is to simply
delete the notification entry.
To that end, we implement a method to delete all
thanks notificaitons.

From 8cf7706a29ded85c07705cfa0
*** Implement requests solved count under avatar ***

From b68194c65df195d54
*** Implement topic bump and others ***

1) Request tag is now only stylized in request sections
2) Bugfix. Firefox not respecting @@username tab completion
3) Add Your topics to quicklinks
4) Chicken bot Easter Egg random event
5) Option for showing topics with most replies
6) Implement Topic bumping
Rules:
Topic bumping is allowed only in listings sections.
Listing section is defined by the root fid in ACP->general settings

In order to bump a topic, first it must be enabled.

A topic bump can be enabled by:
1) Moderator
2) OP && op in allowed_group && topic_bump!=disabled (by mod)

A topic bump can be disabled by:
1) Moderator

Once the topic bump is enabled the topic can be bumped
1) Moderator can bump at any time
2) The OP of the topic can bump once in a cycle defined in ACP

Disabling and enabling topic bump will reset cycle time.

Any topic in listings can be enabled for bumping by moderator,
even if the topic OP is not in the allowed group.