Implement digg
Currently editing the opening post does not send notification.
The only way to send notification is to have users bookmark the topic,
then making a reply to it.
This sends a lot of false positives and is difficult to execute.

With digg, the opening poster can register the topic as digg.
Then any user will be able to digg the topic.
The topic poster will be able to "broadcast" the topic when edits are
made.