Limit quick search maximum amount purchasable
This is a not set in stone number.
But there is a minimum search interval of 5.
And currently, people can buy more search than the quantity required to
bring the interval to a minimum.
Try to limit maximum so that buying the maximum amount will bring the
search interval down to 5 seconds.