Implement django-like RESTful library
Shallow port of django and django-restful-framework library

Problem:
The way pagination works now, it grabs the ENTIRE list of topics.
When the queryset is around 12k, this takes around .3 seconds for each
request. Pagination should not grab everything, it should paginate and
only grab the minimal queryset. I don't expect the topics to get large
(at 1k topics, the query takes around .02 seconds which is acceptable
for now).